Revert "Fixed a few errors with inventories"
This reverts commit 7ee14e54c3c8d09720eaad81397410c99e5abf38.

+++ src/main/java/com/uberverse/arkcraft/common/entity/data/ARKPlayer.java
@@ -209,7 +209,7 @@ public class ARKPlayer implements IExtendedEntityProperties
 	
 	public double getCarryWeightRatio() 
 	{
-		return (double) carryWeight /  (double) weight;
+		return (double) carryWeight / weight;
 	}
 
 	/**
